requirements.txt
anthropic>=0.25.0
requests>=2.31.0
reportlab>=4.1.0
schedule>=1.2.0
python-dotenv>=1.0.0
